25.3.ApiToken Entity{How are Tokens Created? ;Making the ApiToken Class Awesome}
How are Tokens Created?
So, the question of how these ApiTokens will be created is not something we're going to answer. As we talked about, it's either super easy... or super complicated, depending on your needs.
So, for our app, we're just going to create some ApiTokens via the fixtures.

Making the ApiToken Class Awesome
But before we do that, open the new ApiToken entity class. Yep, all the usual stuff: some properties, annotations and a getter & setter for each method. I want to change things a bit. The make:entity command always generates getter and setter methods. But, in some cases, there is a better way to design things.
Add a public function __construct() method with a User argument.
Because every ApiToken needs a User, why not make it required when the object is instantiated? Oh, and we can also generate the random token string here. Use $this->token = bin2hex(random_bytes(60)). Then $this->user = $user.
Oh, and we can also set the expires time here - $this->expiresAt = new \DateTime() with +1 hour.
You can set the expiration time for however long you want.
Now that we are initializing everything in the constructor, we can clean up the class: remove all the setter methods.
Yep, our token class is now immutable, which wins us major hipster points. Immutable just means that, once it's instantiated, this object's data can never be changed. Some developers think that making immutable objects like this is super important. I don't fully agree with that. But, it definitely makes sense to be thoughtful about your entity classes. Sometimes having setter methods makes sense. But sometimes, it makes more sense to setup some things in the constructor and remove the setter methods if you don't need them.
Oh, and if, in the future, you want to update the data in this entity - maybe you need to change the expiresAt, it's totally OK to add a new public function to allow that. But, when you do, again, be thoughtful. You could add a public function setExpiresAt(). Or, if all you ever do is re-set the expiresAt to one hour from now, you could instead create a public function renewExpiresAt() that handles that logic for you:

public function renewExpiresAt()
{
    $this->expiresAt = new \DateTime('+1 hour');
}
That method name is more meaningful, and centralizes more control inside the class.
Ok, I'm done with my rant!