26.1.Entry Point: Helping Users Authenticate{make:auth ApiTokenAuthenticator}
We now have a database table full of API Tokens where each is related to a User. I can already feel the API power! So here's our new goal: when an API request sends a valid API token string, we'll read it and authenticate that request as the User who owns the token.

make:auth ApiTokenAuthenticator
This will be the second way that users can authenticate in our app. So, we need a second authenticator. Find your terminal and run:

php bin/console make:auth
If you see a question about choosing which type of authentication you want, choose an "Empty authenticator". I'm using an older version of the command, which only generates empty authenticators. Call it ApiTokenAuthenticator. Oh, and you may also be asked a question about an "Entry point". We'll talk about that soon, but choose the LoginFormAuthenticator option.
Ok, go check this out!
Hey! I know this class! It's that same, big, adorable empty authenticator we saw earlier. To tell Symfony to use this, open config/packages/security.yaml and add the new class under authenticators.
If you're using that newer, fancier version of this command, it already did this for you. Lucky you!
As soon as we do this, the supports() method will be called at the beginning of every request. But... refresh. Woh! Big error!
  Because you have multiple guard authenticators, you need to set the "guard.entry_point" key to one of your authenticators.
If you did not see this error, it's your lucky day! Well, really, it's because the newer make:auth command took care of this step for you!