27.3.API Token Authenticator{getUser()}
getUser()
Next up: getUser(). First, we need to query for the ApiToken entity. At the top of this class, make an __construct function and give it an ApiTokenRepository  argument. I'll hit Alt+Enter to initialize that.
Then, back in getUser(), get that token: $token = $this->apiTokenRepo->findOneBy() to query where the token property is set to the  string.
If we do not find an ApiToken, return null. That will make authentication fail. If we do find one, we need to return the User, not the token. So, return $token->getUser().
Finally, if you return a User object from getUser(), Symfony calls checkCredentials(). Let's dd('checking credentials') to see if we continue to be lucky.
Move back over to Postman, Send and... yes! Checking credentials.
We're almost done! But before we handle success, I want to see what happens with a bad API key. And learn how we can send back the perfect error response.