28.2.API Token Authenticator Part 2! {onAuthenticationFailure()}
onAuthenticationFailure()
If an API client sends a bad API token, we need to tell them! Bad API client! Let's return a new JsonResponse() with a message key that describes what went wrong. Earlier, I mentioned that whenever authentication fails - for any reason - it's because, internally, some sort of AuthenticationException is thrown. That's important because this exception is passed to us as an argument.
And it has a method - getMessageKey() - that holds a message about what went wrong. Set the status code to 401.