28.3.API Token Authenticator Part 2! {Custom Error Messages with CustomUserMessageAuthenticationException}
Custom Error Messages with CustomUserMessageAuthenticationException
Let's try it again! Send the request. Yes! A 401 Unauthorized response. But, oh. That message isn't right at all!

  Username could not be found?

This is because Symfony creates a different error message based on where authentication fails inside your authenticator. If you fail to return a User from getUser(), you get this "Username could not be found" error.
For our login form, we render this exact messageKey field in the template. But we also pass it through the translator:

 37 lines  templates/security/login.html.twig
{% block body %}
    <form class="form-signin" method="post">
        {% if error %}
            <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
        {% endif %}
    </form>
{% endblock %}
That allowed us to translate that into a better message:

1 lines  translations/security.en.yaml
"Username could not be found.": "Oh no! It doesn't look like that email exists! "
We could do the same here: inject the translator service into ApiTokenAuthenticator and translate the message key. But... hmmm, the message still wouldn't be right - it would use the "It doesn't look like that email exists! " message from the translation file.
No problem: there is a second way to control error messages in an authenticator, and it's super flexible. At any point in your authenticator, you can throw a new CustomUserMessageAuthenticationException() that will cause authentication to fail and accepts any custom error message you want, like, "Invalid API Token"
That's it! This exception will be passed to onAuthenticationFailure() and its getMessageKey() method will return that message.
Go back to Postman to try it: send! We got it! So much better!