28.5.API Token Authenticator Part 2! {onAuthenticationSuccess()}
onAuthenticationSuccess()
At this point... we're basically done! In checkCredentials(), there is no password to check. And so, it's perfectly ok for us to return true.
Finally, in onAuthenticationSuccess(), hmm. What should we do when authentication is successful? With a login form, we redirect the user after success. But with an API token system we, well, want to do... nothing! Yep! We want to allow the request to continue so that it can hit the controller and return the JSON response.