29.1.Manual Authentication / Registration{Creating the Registration Form}
Hey! You've made it through almost this entire tutorial! Nice work! I have just a few more tricks to show you before we're done - and they're good ones!

Creating the Registration Form
First, I want to create a registration form. Find your code and open SecurityController. In addition to login and logout, add a new public function register().
Give it a route - /register and a name: app_register.
Here's the interesting thing about registration. It has nothing to do with security! Think about it. What is registration? It's just a form that creates a new record in the User table. That's it! That's just database stuff.
So then... why are we even talking about this in a security tutorial? Well... to create the best user experience, there will be just a little bit of security right at the end. Because, after registration, I want to instantly authenticate the new user.
More on that later. Right now, render a template: $this->render('security/register.html.twig'). Then... I'll cheat: in security/, copy the login.html.twig template, paste and call it register.html.twig.
Let's see: change the title, delete the authentication error stuff and I am going to add a little comment here that says that we should replace this with a Symfony form later.
We haven't talked about the form system yet, so I don't want to use it here. But, normally, I would use the form system because it handles validation and automatically adds CSRF protection.
But, to show off how to manually authenticate a user after registration, this HTML form will work beautifully. Change the h1, remove the value= on the email field so that it always starts blank and take out the CSRF token.
We do need CSRF protection on this form... but I'll skip it for now, because we'll refactor this into a Symfony form in a future tutorial.
And finally, hijack the "remember me" checkbox and turn it into a terms box. We'll say:

  Agree to terms I for sure read
Oh, and update the button: Register.
Let's see how it looks! Move over, go to /register and... got it! Logout, then move back over and open up base.html.twig. Scroll down just a little bit to find the "Login" link. Let's create a second link that points to the new app_register route. Say, "Register".
Move back and check it out. Not bad!