30.2.Author ManyToOne Relation to User{Adding the Relation; Failed Migration}
Adding the Relation
Now, lets re-add author as a relation:

php bin/console make:entity
Update the Article entity and add a new author property. This will be a "relation" to the User entity. For the type, it's another ManyToOne relation: each Article has one User and each User can have many articles. The author property will be required, so make it not nullable. We'll say "yes" to mapping the other side of the relationship and I'll say "no" to orphanRemoval, though, that's not important. Cool! Hit enter to finish.
Now run:

php bin/console make:migration
Like always, let's go check out the new migration. It adds the author_id column and foreign key constraint. Now run:

php bin/console doctrine:migrations:migrate

Failed Migration!
Woh! It explodes! Bad luck! This is one of those tricky migrations. We made the new column required... but that field will be empty for all the existing rows in the table. That's not a problem on its own... but it does cause a problem when the migration tries to add the foreign key! The fix depends on your situation. If our app were already deployed to production, we would need to follow a 3-step process. First, make the property nullable=true at first and generate that migration. Second, run a script or query that can somehow set the author_id for all the existing articles. And finally, change the property to nullable=false and generate one last migration.
But because our app has not been deployed yet... we can cheat. First, drop all of the tables in the database with:

php bin/console doctrine:schema:drop --full-database --force
Then, re-run all the migrations to make sure they're working:

php bin/console doctrine:migrations:migrate
Awesome! Because the article table is empty, no errors.