31.1.Article Admin & Low-Level Access Controls
Each Article's author is now a proper relationship to the User entity, instead of a string. That's great... except that we haven't updated anything else yet in our code to reflect this. Refresh the homepage. Yep! A big ol' error:

  Exception thrown rendering the template Catchable Fatal Error: Object of Class Proxies\__CG__\App\Entity\User cannot be converted to string.

Wow! Two important things here. First, whenever you see this "Proxies" thing, ignore it. This is an internal object that Doctrine sometimes wraps around your entity in order to enable some of its lazy-loading relation awesomeness. The object looks and works exactly like User.
Second, the error itself basically means that something is trying to convert our User object into a string. This makes sense: in our template, we're just rendering {{ article.author }}:
65 lines  templates/article/homepage.html.twig
{% block body %}
    <div class="container">
        <div class="row">
            <div class="col-sm-12 col-md-8">
                {% for article in articles %}
                <div class="article-container my-1">
                    <a href="{{ path('article_show', {slug: article.slug}) }}">
                        <div class="article-title d-inline-block pl-3 align-middle">
                            <span class="align-left article-details"><img class="article-author-img rounded-circle" src="{{ asset('images/alien-profile.png') }}"> {{ article.author }} </span>
                        </div>
                    </a>
                </div>
                {% endfor %}
            </div>
        </div>
    </div>
{% endblock %}
That was a string before, but now it's a User object.
We could go change this to article.author.firstName. Or, we can go into our User class and add a public function __toString() method. Return $this->getFirstName().
As soon as we do that... we're back!