31.2.Article Admin & Low-Level Access Controls{Adding the Edit Endpoint; Planning the Access Controls}
Adding the Edit Endpoint
What I really want to talk about is controlling access in your system on an object-by-object basis. Like, User A can edit this Article because they are the author, but not that other Article. Open ArticleAdminController and add a new endpoint: public function edit().
Add the normal route with a URL of /admin/article/{id}/edit. I won't give it a name yet. Next, add an argument to the method: Article $article.
Because Article is an entity, SensioFrameworkExtraBundle - a bundle we installed a long time ago - will use the {id} route parameter to query for the correct Article.
To see if this is working, dd($article). Oh, and remember: this entire controller class is protected by ROLE_ADMIN_ARTICLE.
To get a valid Article ID, find your terminal and run:

php bin/console doctrine:query:sql 'SELECT * FROM article'
Ok - we'll use 20. Fly over to you browser and... hit it: /admin/article/20/edit. That bounces us to the login page. Use an admin user: admin2@thespacebar.com password engage.
Perfect! We're back on the Article edit page, access is granted and Doctrine queried for the Article object.

Planning the Access Controls
And this is where things get interesting. I want to continue to require ROLE_ADMIN_ARTICLE to be able to go to the new article page. But, down here, if you're editing an article, I want to allow access if you have ROLE_ADMIN_ARTICLE or if you are the author of this Article. This is the first time that we've had to make an access decision that is based on an object - the Article.