33.2.Adding a Custom Voter{supports(); voteOnAttribute()}
supports()
The  argument will be the string passed to isGranted() and $subject is the second argument - the Article object for us. The example in the generated code is actually pretty good. Let's say that our voter knows how to vote if the $attribute is MANAGE and if the $subject is an instanceOf Article.
If we return false from supports, nothing happens: Our ArticleVoter doesn't vote and it's up to some other voter to handle things. But if we return true, Symfony immediately calls voteOnAttribute().
This is where our logic goes to determine access. If we return true, access will be granted. If we return false, access will be denied.

voteOnAttribute()
Symfony passes us the same $attribute and , as well as something called the $token. The token is a lower-level object that you don't see too often. But, you can use it to get access to the User object.
I'm going to start in this method by helping my editor. At the top, add /** @var Article $subject */ to say that the $subject variable is an Article object.
We can safely do this because of the supports() method. $subject will definitely be an Article at this point.
Below this, it's pretty common to have a voter that votes on multiple attributes, like EDIT and DELETE. We don't need it, but I'll keep the switch case statement. Our only case is MANAGE.
Excellent! It's time to shine. First, if $subject->getAuthor() == $user then return true.
The current user is the author and so access should be granted.